+++ lib/avro2kafka.rb
@@ -3,26 +3,44 @@ require 'avro2kafka/avro_reader'
 require 'avro2kafka/kafka_publisher'
 
 class Avro2Kafka
-  attr_reader :input_path, :kafka_broker, :kafka_topic, :kafka_key
+  attr_reader :options
 
   def initialize(options)
-    @input_path = ARGV.first
-    @kafka_broker = options.delete(:broker)
-    @kafka_topic = options.delete(:topic)
-    @kafka_key = options[:key]
-
     @options = options
   end
 
   def reader
-    ARGF.lineno = 0
-    ARGF
+    ARGF.tap { |argf| argf.rewind }
   end
 
   def publish
     records = AvroReader.new(reader).read
-    KafkaPublisher.new(kafka_broker, kafka_topic, kafka_key).publish(records)
-    puts "Avro file published to #{kafka_topic} topic on #{kafka_broker}!"
+    KafkaPublisher.new(**kafka_options).publish(records)
+    puts "Avro file published to #{topic} topic on #{broker_list}!"
+  end
+
+  def topic
+    options.fetch(:topic)
+  end
+
+  def broker_list
+    options.fetch(:broker_list).split(',')
+  end
+
+  def extra_data
+    options.fetch(:data, []).each_with_object({}) do |data, hash|
+      key, value = data.split('=')
+      hash[key] = value
+    end
+  end
+
+  def kafka_options
+    {
+      broker_list: broker_list,
+      topic: topic,
+      data: extra_data,
+      keys: options.fetch(:key, '').split(',').map(&:strip),
+    }
   end
 
 end